change to floats for speeds

+++ lib/dark_worlds_server/communication/messages.pb.ex
@@ -561,7 +561,7 @@ defmodule DarkWorldsServer.Communication.Proto.Projectile do
   field(:id, 1, type: :uint64)
   field(:position, 2, type: DarkWorldsServer.Communication.Proto.Position)
   field(:direction, 3, type: DarkWorldsServer.Communication.Proto.RelativePosition)
-  field(:speed, 4, type: :uint32)
+  field(:speed, 4, type: :float)
   field(:range, 5, type: :uint32)
   field(:player_id, 6, type: :uint64, json_name: "playerId")
   field(:damage, 7, type: :uint32)
@@ -697,7 +697,7 @@ defmodule DarkWorldsServer.Communication.Proto.GameProjectile do
 
   field(:name, 1, type: :string)
   field(:base_damage, 2, type: :uint64, json_name: "baseDamage")
-  field(:base_speed, 3, type: :uint64, json_name: "baseSpeed")
+  field(:base_speed, 3, type: :float, json_name: "baseSpeed")
   field(:base_size, 4, type: :uint64, json_name: "baseSize")
   field(:remove_on_collision, 5, type: :bool, json_name: "removeOnCollision")
 
@@ -707,7 +707,7 @@ defmodule DarkWorldsServer.Communication.Proto.GameProjectile do
     json_name: "onHitEffect"
   )
 
-  field(:max_distance, 7, type: :uint64, json_name: "maxDistance")
+  field(:max_distance, 7, type: :float, json_name: "maxDistance")
   field(:duration_ms, 8, type: :float, json_name: "durationMs")
 
   def transform_module(), do: DarkWorldsServer.Communication.ProtoTransform
@@ -731,7 +731,7 @@ defmodule DarkWorldsServer.Communication.Proto.GameCharacter do
 
   field(:name, 1, type: :string)
   field(:active, 2, type: :bool)
-  field(:base_speed, 3, type: :uint64, json_name: "baseSpeed")
+  field(:base_speed, 3, type: :float, json_name: "baseSpeed")
   field(:base_size, 4, type: :uint64, json_name: "baseSize")
   field(:base_health, 5, type: :uint64, json_name: "baseHealth")
